+++ R/learner_RWeka_classif_kstar.R
@@ -1,0 +1,83 @@
+#' @title Classification KStar Learner
+#' @author damirpolat
+#' @name mlr_learners_classif.kstar
+#'
+#' @description
+#' Instance-based classifier which differs from other instance-based learners in that
+#' it uses an entropy-based distance function.
+#' Calls [RWeka::make_Weka_classifier()] from \CRANpkg{RWeka}.
+#'
+#' @section Custom mlr3 parameters:
+#' - `output_debug_info`:
+#'   - original id: output-debug-info
+#'
+#' - `do_not_check_capabilities`:
+#'   - original id: do-not-check-capabilities
+#'
+#' - `num_decimal_places`:
+#'   - original id: num-decimal-places
+#'
+#' - `batch_size`:
+#'   - original id: batch-size
+#'
+#' - Reason for change: This learner contains changed ids of the following control arguments
+#' since their ids contain irregular pattern
+#'
+#'
+#' @templateVar id classif.kstar
+#' @template learner
+#'
+#' @references
+#' `r format_bib("Cleary1995")`
+#'
+#' @template seealso_learner
+#' @template example
+#' @export
+LearnerClassifKStar = R6Class("LearnerClassifKStar",
+  inherit = LearnerClassif,
+  public = list(
+    #' @description
+    #' Creates a new instance of this [R6][R6::R6Class] class.
+    initialize = function() {
+      param_set = ps(
+        subset = p_uty(tags = "train"),
+        na.action = p_uty(tags = "train"),
+        B = p_int(default = 20, tags = "train"),
+        E = p_lgl(tags = "train"),
+        M = p_fct(default = "a", levels = c("a", "d", "m", "n"), tags = "train"),
+        output_debug_info = p_lgl(default = FALSE, tags = "train"),
+        do_not_check_capabilities = p_lgl(default = FALSE, tags = "train"),
+        num_decimal_places = p_int(default = 2L, lower = 1L, tags = "train"),
+        batch_size = p_int(default = 100L, lower = 1L, tags = "train"),
+        options = p_uty(default = NULL, tags = "train")
+      )
+
+      super$initialize(
+        id = "classif.kstar",
+        packages = "RWeka",
+        feature_types = c("logical", "integer", "numeric", "factor", "ordered"),
+        predict_types = c("response", "prob"),
+        param_set = param_set,
+        properties = c("missings", "multiclass", "twoclass"),
+        man = "mlr3extralearners::mlr_learners_classif.kstar",
+        label = "KStar"
+      )
+    }
+  ),
+
+  private = list(
+    .train = function(task) {
+      weka_learner = RWeka::make_Weka_classifier("weka/classifiers/lazy/KStar")
+      pars = self$param_set$get_values(tags = "train")
+      rweka_train(task$data(), task$formula(), pars, weka_learner)
+    },
+
+    .predict = function(task) {
+      pars = self$param_set$get_values(tags = "predict")
+      newdata = ordered_features(task, self)
+      rweka_predict(newdata, pars, self$predict_type, self$model)
+    }
+  )
+)
+
+.extralrns_dict$add("classif.kstar", LearnerClassifKStar)